refactor(server): enforce R05 route service boundaries
