+++ crates/server/src/routes/events.rs
@@ -7,6 +7,7 @@ use axum::http::{HeaderMap, StatusCode};
 use axum::response::Response;
 use axum::routing::get;
 use axum::Router;
+use tokio::time::timeout;
 use uuid::Uuid;
 
 use crate::api::sse::{
@@ -15,8 +16,6 @@ use crate::api::sse::{
 };
 use crate::api::{ApiRejection, AppPath};
 use crate::auth::middleware::AuthenticatedOrg;
-use crate::db::pool::with_org_txn;
-use crate::db::sse_streams::{self, ClosedSnapshotLoad, DEFAULT_CLEANUP_LIMIT};
 use crate::http::AppState;
 use crate::routes::common::{deny_or_not_found, map_db};
 use crate::routes::qa_common::{
@@ -24,7 +23,7 @@ use crate::routes::qa_common::{
     revalidate_stream_scope,
 };
 use crate::services::qa::stream::{AuthProbeDecision, StreamCancel};
-use tokio::time::timeout;
+use crate::services::sse_stream::{self, ClosedSnapshotLoad};
 
 pub fn router() -> Router<Arc<AppState>> {
     Router::new().route("/api/v1/events/{request_id}", get(resume_events))
@@ -51,42 +50,18 @@ async fn resume_events(
     )
     .await?;
 
-    let loaded = with_org_txn(state.pool(), &ctx, {
-        let ctx = ctx.clone();
-        move |txn| {
-            Box::pin(
-                async move { sse_streams::load_owned_closed_request(txn, &ctx, request_id).await },
-            )
-        }
-    })
-    .await
-    .map_err(|error| match error {
-        crate::db::error::DbError::NotFound => deny_or_not_found(&http_request_id),
-        other => map_db(other, &http_request_id),
-    })?;
+    let loaded = sse_stream::load_owned_closed_snapshot(state.pool(), &ctx, request_id)
+        .await
+        .map_err(|error| match error {
+            crate::db::error::DbError::NotFound => deny_or_not_found(&http_request_id),
+            other => map_db(other, &http_request_id),
+        })?;
 
     let stream_meta = match loaded {
         ClosedSnapshotLoad::Live(meta) => *meta,
         ClosedSnapshotLoad::Expired { request_id } => {
             // Deterministic 410 first; cleanup after (do not pre-delete before GET).
-            let _ = with_org_txn(state.pool(), &ctx, {
-                let ctx = ctx.clone();
-                move |txn| {
-                    Box::pin(
-                        async move { sse_streams::expire_and_delete(txn, &ctx, request_id).await },
-                    )
-                }
-            })
-            .await;
-            let _ = with_org_txn(state.pool(), &ctx, {
-                let ctx = ctx.clone();
-                move |txn| {
-                    Box::pin(async move {
-                        sse_streams::cleanup_expired(txn, &ctx, DEFAULT_CLEANUP_LIMIT, None).await
-                    })
-                }
-            })
-            .await;
+            let _ = sse_stream::expire_stream_and_cleanup(state.pool(), &ctx, request_id).await;
             return Err(ApiRejection::new(
                 StatusCode::GONE,
                 "stream_expired",
@@ -105,53 +80,28 @@ async fn resume_events(
 
     // Fresh permission + collection scope revalidation before any payload.
     revalidate_stream_scope(&ctx, &stream_meta.auth_scope, &http_request_id)?;
-    let mut probe = make_auth_probe(
-        state.clone(),
-        auth.claims.clone(),
-        stream_meta.auth_scope.clone(),
-    );
+    let mut probe = make_auth_probe(&state, auth.claims.clone(), stream_meta.auth_scope.clone());
     match timeout(SSE_AUTH_PROBE_TIMEOUT, probe()).await {
         Ok(AuthProbeDecision::Allow) => {}
         Ok(other) => return Err(probe_rejection(other, http_request_id)),
         Err(_) => return Err(probe_rejection(AuthProbeDecision::Deny, http_request_id)),
     }
 
-    let events = with_org_txn(state.pool(), &ctx, {
-        let ctx = ctx.clone();
-        move |txn| {
-            Box::pin(async move {
-                sse_streams::list_events_after(txn, &ctx, request_id, after_seq).await
-            })
-        }
-    })
-    .await
-    .map_err(|error| match error {
-        crate::db::error::DbError::NotFound => deny_or_not_found(&http_request_id),
-        other => map_db(other, &http_request_id),
-    })?;
-
-    // Bounded opportunistic cleanup; never remove the active request.
-    let _ = with_org_txn(state.pool(), &ctx, {
-        let ctx = ctx.clone();
-        move |txn| {
-            Box::pin(async move {
-                sse_streams::cleanup_expired(txn, &ctx, DEFAULT_CLEANUP_LIMIT, Some(request_id))
-                    .await
-            })
-        }
-    })
-    .await;
+    let events =
+        sse_stream::list_events_after_and_cleanup(state.pool(), &ctx, request_id, after_seq)
+            .await
+            .map_err(|error| match error {
+                crate::db::error::DbError::NotFound => deny_or_not_found(&http_request_id),
+                other => map_db(other, &http_request_id),
+            })?;
 
     let envelopes: Vec<_> = events
         .iter()
         .map(|row| event_row_to_envelope(row, &request_id.to_string()))
         .collect();
     let cancel = StreamCancel::new();
-    let delivery_probe = make_auth_probe(
-        state.clone(),
-        auth.claims.clone(),
-        stream_meta.auth_scope.clone(),
-    );
+    let delivery_probe =
+        make_auth_probe(&state, auth.claims.clone(), stream_meta.auth_scope.clone());
     let stream = deliver_closed_snapshot(
         envelopes,
         cancel,
